Email Confirmation
Added an automated email which will be sent after someone enters their
information. Changed the preferred scheme so that it is now storing the
label not the value.
Added some borders to field sets to make it look nicer.
Included a quizResults page which will be the resulting page after
submitting your quiz answers. This page still needs to be styled.